Moved console related classes to new package
